implementing request manager
