+++ projects/sw-background-sync-queue/src/lib/idbHelper.js
@@ -1,0 +1,11 @@
+import IDBHelper from '../../../../lib/idb-helper';
+
+let idbQueue;
+
+(async function idbHelper(){
+	let idbQHelper = new IDBHelper("bgQueueSyncDB",1,"QueueStore");
+	idbQueue = await idbQHelper.get("queue");
+})();
+
+
+export { idbQueue };
